+++ TransactionEntites/Entities/Trackers/StudentActivityTracker.cs
@@ -5,7 +5,7 @@ namespace TransactionEntites.Entities.Trackers;
 public class StudentActivityTracker : TrackableEntity
 {
     public Guid StudentId { get; set; }
-    public int StudentPoints { get; set; }
+    public double StudentPoints { get; set; }
     public int LearningDurationInSec { get; set; }
     public bool IsActive { get; set; }
     public int Code { get; set; }
